requirements.txt
# general  modules
django>3.2,<3.3
# python-decouple
# Pillow
djangorestframework
psycopg2-binary

# third party modules
#markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
#djoser
django-cors-headers

# deployment module
# gunicorn

# email third party modules
#django-mail-templated

# refactoring and tests
# black
# flake8
#
pytest
pytest-django
#Faker

# background proccess and cache
# celery
# redis
# django-redis
# django-celery-beat